[rocprofiler-compute] Consolidate native PC sampling tests
- Merge the three native test files (ISA map loading, enrichment, and
  load_pc_sampling_data threading) into one test_pc_sampling_native.py
  with single-copy helpers and one-line docstrings
- Drop two enrich_with_metadata basis tests duplicated from the
  consolidated enrichment cases
